Split TransformHandles into two mods
